Fixed double float conflict between SqlServer & ASP.Net

+++ WebService.API/Datas/Models/Products/Vga.cs
@@ -4,7 +4,7 @@ namespace WebService.API.Datas.Models.Products;
 
 public record Vga : ModelBase
 {
-	public float Frequency { get; set; }
+	public double Frequency { get; set; }
 	public int Vram { get; set; }
 	public virtual Product? Product { get; set; }
 }
